major edits

+++ src/main/java/at/fh/controller/MediaController.java
@@ -25,11 +25,13 @@ public class MediaController extends BaseController implements HttpHandler {
     // ROUTING
     @Override
     public void handle(HttpExchange ex) throws IOException {
-        String path = ex.getRequestURI().getPath();
+        String fullPath = ex.getRequestURI().getPath();
+        String context  = ex.getHttpContext().getPath();
+        String path = fullPath.substring(context.length());
         String method = ex.getRequestMethod();
 
         try {
-            if (path.equals("/api/media/") || path.equals("/api/media")) {
+            if (path.equals("/") || path.isEmpty()) {
                 switch (method) {
                     case "GET":
                         findAllMedia(ex);
@@ -41,34 +43,37 @@ public class MediaController extends BaseController implements HttpHandler {
                         sendText(ex, 405, "Not a valid method");
                         return;
                 }
-            }
-
-            if (path.startsWith("/api/media/")) {
-                String id = path.substring("/api/media/".length());
-                UUID mediaId = UUID.fromString(id);
-
+            } else if ((path.matches("^/[^/]+$"))) {
                 switch (method) {
                     case "GET":
-                        findMediaById(ex,mediaId);
+                        findMediaById(ex);
                         return;
                     case "PUT":
-                        updateMedia(ex,mediaId);
+                        updateMedia(ex);
                         return;
                     case "DELETE":
-                        deleteMedia(ex,mediaId);
+                        deleteMedia(ex);
                         return;
                     default:
                         sendText(ex, 405, "Not a valid method");
                 }
+                } else if (path.matches("^/[^/]+/rate/?$")) {
+                    //TODO: FERTIG MACHEN
+            } else {
+                sendText(ex, 404, "Endpoint not found");
+                return;
             }
         } catch (Exception e) {
             sendText(ex, 500, "Internal Server Error");
             return;
         }
 
-    }
+    } //TODO: MEDIARATING FERTIG MACHEN
 
     private void findAllMedia(HttpExchange ex) throws IOException {
+        UUID loggedInUserId = authorizeUser(ex);
+        if (loggedInUserId == null) return;
+
         List<MediaEntry> result = mediaService.findAll();
 
         if (result.isEmpty()){
@@ -82,6 +87,8 @@ public class MediaController extends BaseController implements HttpHandler {
 
     private void createMedia(HttpExchange ex) throws IOException {
         UUID userId = authorizeUser(ex);
+        if (userId == null) return;
+
         //Liest die Daten, die der User über JSON mitgegeben hat!
         MediaInput req = readJson(ex, MediaInput.class);
 
@@ -91,21 +98,28 @@ public class MediaController extends BaseController implements HttpHandler {
             return;
         }
 
-        ex.sendResponseHeaders(201, -1);
-        ex.close();
+        sendText(ex, 201, "Media created");
     }
 
-    private void findMediaById(HttpExchange ex, UUID mediaId) throws IOException {
+    private void findMediaById(HttpExchange ex) throws IOException {
+        UUID userId = authorizeUser(ex);
+        if (userId == null) return;
+
+        UUID mediaId = extractUuid(ex);
+
         Optional<MediaEntry> result = mediaService.findById(mediaId);
         if (result.isEmpty()) {
             sendText(ex, 404, "Media not found");
             return;
         }
-        ex.sendResponseHeaders(200, -1);
-        ex.close();
+        sendJson(ex, 200, result);
     }
 
-    private void updateMedia(HttpExchange ex, UUID mediaId) throws IOException {
+    private void updateMedia(HttpExchange ex) throws IOException {
+        UUID userId = authorizeUser(ex);
+        if (userId == null) return;
+
+        UUID mediaId = extractUuid(ex);
 
         //Liest die Daten, die der User über JSON mitgegeben hat!
         MediaInput req = readJson(ex, MediaInput.class);
@@ -119,7 +133,12 @@ public class MediaController extends BaseController implements HttpHandler {
         ex.close();
     }
 
-    private void deleteMedia(HttpExchange ex, UUID mediaId) throws IOException {
+    private void deleteMedia(HttpExchange ex) throws IOException {
+        UUID userId = authorizeUser(ex);
+        if (userId == null) return;
+
+        UUID mediaId = extractUuid(ex);
+
         boolean success = mediaService.delete(mediaId);
 
         if (!success) {
